requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JSGlue==0.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.3
future==0.18.2
idna==2.9
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.2
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.0
python-socketio==4.6.0
PyYAML==5.3.1
requests==2.23.0
serial==0.0.97
six==1.15.0
SQLAlchemy==1.3.17
sqlalchemy-json==0.4.0
urllib3==1.25.9
Werkzeug==1.0.1